Run setupEventsManager in the LandingPresenter
We can only grab events if the user has a valid JWT, which we can't
get unless the user is authenticated. Thus it doesn't make sense
for us to run setupEventsManager in the SplashPresenter in case the
user is not logged in. So, let's run it at a place where we're
guarenteed that the user is authenticated (in our case, when they
reach the LandingPresenter).

We employ an observer pattern here so that components which rely on
events can be notified about loading and finished loading states.